Update train.py to avoid attribute errors
As explained in detail in a [numpy issue](https://github.com/numpy/numpy/issues/4701), `np.asscalar` doesn't pass through scalars. This is because it looks for `item()` method that `int` and `float` don't have. A possible workaround suggest there is the one I proposed to you. In this way your script doesn't fail with other environments such as `LunarLanderContinuous-v2`.